requirements.txt
Pillow>=10
pillow-heif>=0.13
tkinterdnd2
